+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Base class for excluded parameters.
+- Resolving of asset/hierarchy id within tools.
+
 ## [v0.11.2] - 3.11.2025
 
 ### Fixed